+++ lib/subvisual/array_utils.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "subvisual/array_utils/pad"
+
+module Subvisual
+  module ArrayUtils
+    extend Pad
+  end
+end