+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning].
 Unreleased
 ----------
 
-_Coming soon_
+- Added `ArrayUtils.pad_right(array, size, padding = nil)`.
 
 
 v0.1.0 - 2020-05-06